+++ WalletWasabi/News/README.md
@@ -1,3 +1,3 @@
 `ClientUpdates.json` is a file that must be manually copied to the `.walletwasabi\backend` directory whenever new item is added. Clients will synchronize it down and show it on their dashboards.
 
-`cp ~/WalletWasabi/WalletWasabi/ClientUpdates/ClientUpdates.json ~/.walletwasabi/backend/ClientUpdates.json`
+`cp ~/WalletWasabi/WalletWasabi/News/ClientUpdates.json ~/.walletwasabi/backend/ClientUpdates.json`
